光滑元例子修改

+++ fealpy/mesh/triangle_mesh.py
@@ -1110,7 +1110,7 @@ class TriangleMesh(SimplexMesh, Plotable):
         @return TriangleMesh instance
         """
         if dtype is None:
-            dtype = bm.int32
+            dtype = bm.float64
 
         NN = (nx + 1) * (ny + 1)
         NC = nx * ny
